ai-privacy-guard/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-privacy-guard"
version = "0.1.0"
description = "Policy-driven privacy risk checks for AI workloads."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "AI Privacy Guard Contributors" }
]
dependencies = [
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
apg = "ai_privacy_guard.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
